move tests to separate classes
