+++ core/src/test/scala/org/broadinstitute/dsde/rawls/workspace/MultiCloudWorkspaceServiceCloneSpec.scala
@@ -1,0 +1,15 @@
+package org.broadinstitute.dsde.rawls.workspace
+
+import org.scalatest.OptionValues
+import org.scalatest.concurrent.ScalaFutures
+import org.scalatest.flatspec.AnyFlatSpecLike
+import org.scalatest.matchers.should.Matchers
+import org.scalatestplus.mockito.MockitoSugar
+
+class MultiCloudWorkspaceServiceCloneSpec extends AnyFlatSpecLike
+  with MockitoSugar
+  with ScalaFutures
+  with Matchers
+  with OptionValues {
+
+}
