+++ core/src/test/scala/org/broadinstitute/dsde/rawls/workspace/MultiCloudWorkspaceServiceSpec.scala
@@ -13,11 +13,35 @@ import org.broadinstitute.dsde.rawls.config.{AzureConfig, MultiCloudWorkspaceCon
 import org.broadinstitute.dsde.rawls.dataaccess.slick.{TestDriverComponent, WorkspaceManagerResourceMonitorRecord}
 import org.broadinstitute.dsde.rawls.dataaccess.slick.WorkspaceManagerResourceMonitorRecord.JobType
 import org.broadinstitute.dsde.rawls.dataaccess.workspacemanager.WorkspaceManagerDAO
-import org.broadinstitute.dsde.rawls.dataaccess.{LeonardoDAO, MockLeonardoDAO, SamDAO, WorkspaceManagerResourceMonitorRecordDao}
+import org.broadinstitute.dsde.rawls.dataaccess.{
+  LeonardoDAO,
+  MockLeonardoDAO,
+  SamDAO,
+  WorkspaceManagerResourceMonitorRecordDao
+}
 import org.broadinstitute.dsde.rawls.mock.{MockSamDAO, MockWorkspaceManagerDAO}
-import org.broadinstitute.dsde.rawls.model.WorkspaceState.WorkspaceState
 import org.broadinstitute.dsde.rawls.model.WorkspaceType.McWorkspace
-import org.broadinstitute.dsde.rawls.model.{AttributeBoolean, AttributeName, AttributeString, ErrorReport, GoogleProjectId, RawlsBillingProject, RawlsRequestContext, SamBillingProjectActions, SamResourceTypeNames, SamUserStatusResponse, SamWorkspaceActions, Workspace, WorkspaceCloudPlatform, WorkspaceDeletionResult, WorkspaceName, WorkspacePolicy, WorkspaceRequest, WorkspaceState, WorkspaceType, WorkspaceVersions}
+import org.broadinstitute.dsde.rawls.model.{
+  AttributeName,
+  AttributeString,
+  ErrorReport,
+  GoogleProjectId,
+  RawlsBillingProject,
+  RawlsRequestContext,
+  SamBillingProjectActions,
+  SamResourceTypeNames,
+  SamUserStatusResponse,
+  SamWorkspaceActions,
+  Workspace,
+  WorkspaceCloudPlatform,
+  WorkspaceDeletionResult,
+  WorkspaceName,
+  WorkspacePolicy,
+  WorkspaceRequest,
+  WorkspaceState,
+  WorkspaceType,
+  WorkspaceVersions
+}
 import org.broadinstitute.dsde.rawls.workspace.MultiCloudWorkspaceService.getStorageContainerName
 import org.broadinstitute.dsde.workbench.client.leonardo
 import org.joda.time.DateTime
@@ -58,7 +82,6 @@ class MultiCloudWorkspaceServiceSpec
       landingZoneAllowAttach = false
     )
   )
-
 
   behavior of "createMultiCloudWorkspace"
 
@@ -145,8 +168,6 @@ class MultiCloudWorkspaceServiceSpec
         ArgumentMatchers.any()
       )
   }
-
-
 
   behavior of "cloneMultiCloudWorkspace"
 
@@ -745,239 +766,7 @@ class MultiCloudWorkspaceServiceSpec
       }
     }
 
-  behavior of "deleteMultiCloudOrRawlsWorkspaceV2"
-
-  it should "call WorkspaceService to delete a rawls workspace synchronously" in {
-    val namespace = "ws_namespace"
-    val name = "ws_name"
-    val workspaceName = WorkspaceName(namespace, name)
-    val workspaceId = UUID.randomUUID()
-    val samDAO = mock[SamDAO]
-    when(samDAO.getUserStatus(any())).thenReturn(Future(Some(SamUserStatusResponse("", "", true))))
-    when(samDAO.userHasAction(SamResourceTypeNames.workspace, workspaceId.toString,SamWorkspaceActions.delete,testContext))
-      .thenReturn(Future(true))
-    val workspaceRepository = mock[WorkspaceRepository]
-    val workspace = Workspace(
-      namespace,
-      name,
-      workspaceId.toString,
-      "test-bucket-name",
-      None,
-      DateTime.now(),
-      DateTime.now(),
-      "",
-      Map.empty,
-      false,
-      WorkspaceVersions.V2,
-      GoogleProjectId("project-id"),
-      None,
-      None,
-      None,
-      None,
-      WorkspaceType.RawlsWorkspace,
-      WorkspaceState.Ready
-    )
-    when(workspaceRepository.getWorkspace(workspaceName, None)).thenReturn(Future(Some(workspace)))
-    val workspaceService = mock[WorkspaceService]
-    val deleteResult = WorkspaceDeletionResult(None, None)
-    when(workspaceService.deleteWorkspace(workspaceName)).thenReturn(Future(deleteResult))
-    val service = new MultiCloudWorkspaceService(
-      testContext,
-      mock[WorkspaceManagerDAO],
-      mock[BillingProfileManagerDAO],
-      samDAO,
-      mock[MultiCloudWorkspaceConfig],
-      mock[LeonardoDAO],
-      "MultiCloudWorkspaceService-test",
-      mock[WorkspaceManagerResourceMonitorRecordDao],
-      workspaceRepository,
-      mock[BillingRepository]
-    )
-
-    Await.result(
-      service.deleteMultiCloudOrRawlsWorkspaceV2(workspaceName, workspaceService),
-      Duration.Inf
-    ) shouldBe deleteResult
-    verify(workspaceService).deleteWorkspace(workspaceName)
-  }
-
-  it should "create a new job for the workspace manager resource monitor when deleting an azure workspace" in {
-    val namespace = "ws_namespace"
-    val name = "ws_name"
-    val workspaceName = WorkspaceName(namespace, name)
-    val workspaceId = UUID.randomUUID()
-    val samDAO = mock[SamDAO]
-    when(samDAO.getUserStatus(any())).thenReturn(Future(Some(SamUserStatusResponse("", "", true))))
-    when(samDAO.userHasAction(SamResourceTypeNames.workspace, workspaceId.toString, SamWorkspaceActions.delete, testContext))
-      .thenReturn(Future(true))
-    val workspace = Workspace.buildMcWorkspace(
-      namespace = namespace,
-      name = name,
-      workspaceId = workspaceId.toString,
-      DateTime.now(),
-      DateTime.now(),
-      createdBy = testContext.userInfo.userEmail.value,
-      attributes = Map.empty,
-      state = WorkspaceState.Ready
-    )
-    val workspaceRepository = mock[WorkspaceRepository]
-    when(workspaceRepository.getWorkspace(workspaceName, None)).thenReturn(Future(Some(workspace)))
-    when(workspaceRepository.updateState(workspace.workspaceIdAsUUID, WorkspaceState.Deleting)).thenReturn(Future(1))
-    val monitorRecordDao = mock[WorkspaceManagerResourceMonitorRecordDao]
-    val recordCaptor = ArgumentCaptor.forClass(classOf[WorkspaceManagerResourceMonitorRecord])
-    when(monitorRecordDao.create(recordCaptor.capture())).thenReturn(Future())
-
-    val service = new MultiCloudWorkspaceService(
-      testContext,
-      mock[WorkspaceManagerDAO],
-      mock[BillingProfileManagerDAO],
-      samDAO,
-      mock[MultiCloudWorkspaceConfig],
-      mock[LeonardoDAO],
-      "MultiCloudWorkspaceService-test",
-      monitorRecordDao,
-      workspaceRepository,
-      mock[BillingRepository]
-    )
-    val deletionResult = Await.result(
-      service.deleteMultiCloudOrRawlsWorkspaceV2(workspaceName, mock[WorkspaceService]),
-      Duration.Inf
-    )
-    val record: WorkspaceManagerResourceMonitorRecord = recordCaptor.getValue
-    record.workspaceId shouldBe Some(workspaceId)
-    record.jobType shouldBe WorkspaceManagerResourceMonitorRecord.JobType.WorkspaceDeleteInit
-    record.jobControlId shouldBe deletionResult.jobId.map(UUID.fromString).get
-  }
-
-  it should "fail with a conflict if the workspace is already deleting" in {
-    val namespace = "ws_namespace"
-    val name = "ws_name"
-    val workspaceName = WorkspaceName(namespace, name)
-    val workspaceId = UUID.randomUUID()
-    val samDAO = mock[SamDAO]
-    when(samDAO.getUserStatus(any())).thenReturn(Future(Some(SamUserStatusResponse("", "", true))))
-    when(samDAO.userHasAction(SamResourceTypeNames.workspace, workspaceId.toString, SamWorkspaceActions.delete, testContext))
-      .thenReturn(Future(true))
-    val workspace = Workspace.buildMcWorkspace(
-      namespace = namespace,
-      name = name,
-      workspaceId = workspaceId.toString,
-      DateTime.now(),
-      DateTime.now(),
-      createdBy = testContext.userInfo.userEmail.value,
-      attributes = Map.empty,
-      state = WorkspaceState.Deleting
-    )
-    val workspaceRepository = mock[WorkspaceRepository]
-    when(workspaceRepository.getWorkspace(workspaceName, None)).thenReturn(Future(Some(workspace)))
-    val service = new MultiCloudWorkspaceService(
-      testContext,
-      mock[WorkspaceManagerDAO],
-      mock[BillingProfileManagerDAO],
-      samDAO,
-      mock[MultiCloudWorkspaceConfig],
-      mock[LeonardoDAO],
-      "MultiCloudWorkspaceService-test",
-      mock[WorkspaceManagerResourceMonitorRecordDao],
-      workspaceRepository,
-      mock[BillingRepository]
-    )
-
-    val result = intercept[RawlsExceptionWithErrorReport] {
-       Await.result(service.deleteMultiCloudOrRawlsWorkspaceV2(workspaceName, mock[WorkspaceService]), Duration.Inf)
-     }
-
-    result.errorReport.statusCode shouldBe Some(StatusCodes.Conflict)
-  }
-
-  it should "fail with an error if the workspace is in an undeleteable state" in {
-    val namespace = "ws_namespace"
-    val name = "ws_name"
-    val workspaceName = WorkspaceName(namespace, name)
-    val workspaceId = UUID.randomUUID()
-    val samDAO = mock[SamDAO]
-    when(samDAO.getUserStatus(any())).thenReturn(Future(Some(SamUserStatusResponse("", "", true))))
-    when(samDAO.userHasAction(SamResourceTypeNames.workspace, workspaceId.toString, SamWorkspaceActions.delete, testContext))
-      .thenReturn(Future(true))
-    val workspace = Workspace.buildMcWorkspace(
-      namespace = namespace,
-      name = name,
-      workspaceId = workspaceId.toString,
-      DateTime.now(),
-      DateTime.now(),
-      createdBy = testContext.userInfo.userEmail.value,
-      attributes = Map.empty,
-      state = WorkspaceState.Creating
-    )
-    val workspaceRepository = mock[WorkspaceRepository]
-    when(workspaceRepository.getWorkspace(workspaceName, None)).thenReturn(Future(Some(workspace)))
-    val service = new MultiCloudWorkspaceService(
-      testContext,
-      mock[WorkspaceManagerDAO],
-      mock[BillingProfileManagerDAO],
-      samDAO,
-      mock[MultiCloudWorkspaceConfig],
-      mock[LeonardoDAO],
-      "MultiCloudWorkspaceService-test",
-      mock[WorkspaceManagerResourceMonitorRecordDao],
-      workspaceRepository,
-      mock[BillingRepository]
-    )
-
-    val result = intercept[RawlsExceptionWithErrorReport] {
-      Await.result(service.deleteMultiCloudOrRawlsWorkspaceV2(workspaceName, mock[WorkspaceService]), Duration.Inf)
-    }
-
-    result.errorReport.statusCode shouldBe Some(StatusCodes.BadRequest)
-  }
-
   behavior of "deleteMultiCloudOrRawlsWorkspace"
-
-  it should "fail if user does not have delete permissions on the workspace" in {
-    val namespace = "ws_namespace"
-    val name = "ws_name"
-    val workspaceName = WorkspaceName(namespace, name)
-    val workspaceId = UUID.randomUUID()
-    val samDAO = mock[SamDAO]
-    when(samDAO.getUserStatus(any())).thenReturn(Future(Some(SamUserStatusResponse("", "", true))))
-    when(samDAO.userHasAction(SamResourceTypeNames.workspace, workspaceId.toString, SamWorkspaceActions.delete, testContext))
-      .thenReturn(Future(false))
-    when(samDAO.userHasAction(SamResourceTypeNames.workspace, workspaceId.toString, SamWorkspaceActions.read, testContext))
-      .thenReturn(Future(true))
-    val workspace = Workspace.buildMcWorkspace(
-      namespace = namespace,
-      name = name,
-      workspaceId = workspaceId.toString,
-      DateTime.now(),
-      DateTime.now(),
-      createdBy = testContext.userInfo.userEmail.value,
-      attributes = Map.empty,
-      state = WorkspaceState.Creating
-    )
-    val workspaceRepository = mock[WorkspaceRepository]
-    when(workspaceRepository.getWorkspace(workspaceName, None)).thenReturn(Future(Some(workspace)))
-    val service = new MultiCloudWorkspaceService(
-      testContext,
-      mock[WorkspaceManagerDAO],
-      mock[BillingProfileManagerDAO],
-      samDAO,
-      mock[MultiCloudWorkspaceConfig],
-      mock[LeonardoDAO],
-      "MultiCloudWorkspaceService-test",
-      mock[WorkspaceManagerResourceMonitorRecordDao],
-      workspaceRepository,
-      mock[BillingRepository]
-    )
-
-    val result = intercept[RawlsExceptionWithErrorReport] {
-      Await.result(
-          service.deleteMultiCloudOrRawlsWorkspace(workspaceName, mock[WorkspaceService])
-,        Duration.Inf
-      )
-    }
-
-    result.errorReport.statusCode shouldEqual Some(StatusCodes.Forbidden)
-  }
 
   it should "delete an MC workspace" in {
     withEmptyTestDatabase {
@@ -1036,7 +825,13 @@ class MultiCloudWorkspaceServiceSpec
     val workspaceId = UUID.randomUUID()
     val samDAO = mock[SamDAO]
     when(samDAO.getUserStatus(any())).thenReturn(Future(Some(SamUserStatusResponse("", "", true))))
-    when(samDAO.userHasAction(SamResourceTypeNames.workspace, workspaceId.toString, SamWorkspaceActions.delete, testContext))
+    when(
+      samDAO.userHasAction(SamResourceTypeNames.workspace,
+                           workspaceId.toString,
+                           SamWorkspaceActions.delete,
+                           testContext
+      )
+    )
       .thenReturn(Future(true))
     val workspace = Workspace.buildMcWorkspace(
       namespace = namespace,
@@ -1051,9 +846,7 @@ class MultiCloudWorkspaceServiceSpec
     val workspaceRepository = mock[WorkspaceRepository]
     when(workspaceRepository.getWorkspace(workspaceName, None)).thenReturn(Future(Some(workspace)))
     val wsmDAO = mock[WorkspaceManagerDAO]
-    when(wsmDAO.getWorkspace(workspaceId, testContext)).thenAnswer(_ =>
-      throw new ApiException(500, "error")
-    )
+    when(wsmDAO.getWorkspace(workspaceId, testContext)).thenAnswer(_ => throw new ApiException(500, "error"))
     val service = new MultiCloudWorkspaceService(
       testContext,
       wsmDAO,
@@ -1084,7 +877,13 @@ class MultiCloudWorkspaceServiceSpec
     val workspaceId = UUID.randomUUID()
     val samDAO = mock[SamDAO]
     when(samDAO.getUserStatus(any())).thenReturn(Future(Some(SamUserStatusResponse("", "", true))))
-    when(samDAO.userHasAction(SamResourceTypeNames.workspace, workspaceId.toString, SamWorkspaceActions.delete, testContext))
+    when(
+      samDAO.userHasAction(SamResourceTypeNames.workspace,
+                           workspaceId.toString,
+                           SamWorkspaceActions.delete,
+                           testContext
+      )
+    )
       .thenReturn(Future(true))
     val workspace = Workspace.buildMcWorkspace(
       namespace = namespace,
@@ -1100,9 +899,7 @@ class MultiCloudWorkspaceServiceSpec
     when(workspaceRepository.getWorkspace(workspaceName, None)).thenReturn(Future(Some(workspace)))
     when(workspaceRepository.deleteWorkspace(workspace)).thenReturn(Future(true))
     val wsmDAO = mock[WorkspaceManagerDAO]
-    when(wsmDAO.getWorkspace(workspaceId, testContext)).thenAnswer(_ =>
-      throw new ApiException(404, "not found")
-    )
+    when(wsmDAO.getWorkspace(workspaceId, testContext)).thenAnswer(_ => throw new ApiException(404, "not found"))
     val service = new MultiCloudWorkspaceService(
       testContext,
       wsmDAO,
@@ -1116,7 +913,8 @@ class MultiCloudWorkspaceServiceSpec
       mock[BillingRepository]
     )
 
-    val result = Await.result(service.deleteMultiCloudOrRawlsWorkspace(workspaceName, mock[WorkspaceService]), Duration.Inf)
+    val result =
+      Await.result(service.deleteMultiCloudOrRawlsWorkspace(workspaceName, mock[WorkspaceService]), Duration.Inf)
     result shouldBe None
 
     verify(wsmDAO).getWorkspace(workspaceId, testContext)
@@ -1320,7 +1118,6 @@ class MultiCloudWorkspaceServiceSpec
     )
   }
 
-
   private def assertWorkspaceGone(workspace: Workspace) = {
     // fail if the workspace exists
     val existing = Await.result(
@@ -1330,7 +1127,6 @@ class MultiCloudWorkspaceServiceSpec
 
     existing shouldBe empty
   }
-
 
   private def insertWorkspaceWithBillingProject(billingProject: RawlsBillingProject, workspace: Workspace) =
     slickDataSource.inTransaction { access =>
